# backend/src/main/resources/api-docs.yaml
openapi: 3.1.0
info:
  title: Lyric Vocabulary Builder API
  description: |-
    英语歌词词汇构建器应用程序的后台 API。
    用于管理英文歌曲库、构建词汇索引并提供英英词典查询，专为英语学习设计。

    Backend API for the Lyric Vocabulary Builder application (English learning only).

    所有包含响应体的接口统一返回 `{ code, message, data }`。本文档中的成功响应 schema
    描述 `data` 字段的业务内容；前端生成客户端会在传输层自动解包该字段。

    All responses with a body use the `{ code, message, data }` envelope. Successful
    response schemas below describe the business payload inside `data`; the generated
    frontend transport unwraps this field automatically.
  version: 1.3.0
servers:
  - url: http://localhost:8080
    description: 本地开发服务器 / Local development server
...
paths:
...
  /api/vocabulary/words:
    get:
      tags: [Vocabulary]
      summary: 获取唯一单词分页列表 / Get paginated unique words
      description: |-
        返回所有英文歌词中提取的唯一单词（已排序、不区分大小写去重），支持前缀过滤和分页。
      operationId: getWordList
      parameters:
        - name: prefix
          in: query
          schema:
            type: string
          description: 前缀过滤（不区分大小写） / Case-insensitive prefix filter
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: 页码（从0开始） / Page number (0-indexed)
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: 每页条数 / Page size
        - name: recommendedOnly
          in: query
          schema:
            type: boolean
            default: true
          description: 是否只返回推荐学习词 / Whether to return recommended words only
        - name: lemmaSearch
          in: query
          schema:
            type: boolean
            default: true
          description: 是否使用词形归一搜索 / Whether to use lemmatized search
      responses:
        '200':
          description: 分页单词列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WordPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Vocabulary]
      summary: 批量删除词汇索引项 / Delete vocabulary index words
      description: 删除词汇索引中的指定词条，不删除歌曲原文。
      operationId: deleteVocabularyWords
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VocabularyBulkWordsRequest'
      responses:
        '200':
          description: 实际删除数量
          content:
            application/json:
              schema:
                type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/vocabulary/words/{word}/occurrences:
    get:
      tags: [Vocabulary]
      summary: 获取单词出现位置 / Get word occurrences in lyrics
      operationId: getWordOccurrences
      parameters:
        - name: word
          in: path
          required: true
          description: 要查询的英文单词（不区分大小写）
          schema:
            type: string
      responses:
        '200':
          description: 出现位置列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WordOccurrence'
        '404':
          description: 词汇表中未找到该单词
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/vocabulary/words/{word}/learning-value:
    patch:
      tags: [Vocabulary]
      summary: 手动更新词条学习价值 / Manually update word learning value
      description: 将词条标记为推荐学习或低学习价值，不删除歌曲原文。
      operationId: updateVocabularyLearningValue
      parameters:
        - name: word
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VocabularyLearningValueRequest'
      responses:
        '200':
          description: 更新后的词条摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VocabularyQualityCandidate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/vocabulary/quality-candidates:
    get:
      tags: [Vocabulary]
      summary: 获取词库清洗候选 / Get vocabulary cleanup candidates
      description: 返回疑似低学习价值或异常词条，用于人工确认后清洗。
      operationId: getVocabularyQualityCandidates
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 80
      responses:
        '200':
          description: 词库清洗候选
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VocabularyQualityCandidate'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/vocabulary/refresh:
    post:
      tags: [Vocabulary]
      summary: 重新构建词汇索引 / Rebuild vocabulary index
      description: 触发异步任务重新从所有歌曲提取词汇索引
      operationId: refreshVocabularyIndex
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '202':
          description: 异步重建任务已接受
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreationResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/vocabulary/refresh/tasks/{taskId}:
    get:
      tags: [Vocabulary]
      summary: 查询词汇索引刷新任务状态
      description: 根据任务 ID 查询词汇索引重建任务的当前进度和最终结果。
      operationId: getRefreshTaskStatus
      parameters:
        - name: taskId
          in: path
          required: true
          description: 从 /refresh 接口获取的任务 ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 任务的当前状态或最终结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VocabularyRebuildTask'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/vocabulary/phrases:
    get:
      tags: [Phrases]
      summary: 分页查询歌曲中实际命中的正式短语 / Get paginated phrases matched in songs
      operationId: listPhrases
      parameters:
        - name: q
          in: query
          schema: {type: string}
          description: 在歌曲实际命中的短语中进行包含匹配 / Contains search among phrases matched in songs
        - name: page
          in: query
          schema: {type: integer, minimum: 0, default: 0}
        - name: size
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
      responses:
        '200':
          description: 分页短语列表
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PhrasePage'}
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/vocabulary/phrases/search:
    get:
      tags: [Phrases]
      summary: 搜索歌曲中实际命中的正式短语 / Search dictionary phrases matched in songs
      operationId: searchPhrases
      parameters:
        - name: q
          in: query
          required: true
          schema: {type: string}
          description: 在歌曲实际命中的短语中进行包含匹配 / Contains search among phrases matched in songs
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
      responses:
        '200':
          description: 短语列表
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/DictionaryPhrase'}
  /api/songs/{songId}/phrases:
    get:
      tags: [Phrases]
      summary: 查询歌曲中的正式短语 / Get phrases in a song
      operationId: getSongPhrases
      parameters: [{ $ref: '#/components/parameters/SongId' }]
      responses:
        '200':
          description: 歌曲短语命中
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/PhraseMatch'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/vocabulary/phrases/{phraseId}/occurrences:
    get:
      tags: [Phrases]
      summary: 查询短语出现位置 / Get phrase occurrences
      operationId: getPhraseOccurrences
      parameters:
        - name: phraseId
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        '200':
          description: 短语出现位置列表
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/PhraseOccurrence'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/songs/{songId}/phrases/refresh:
    post:
      tags: [Phrases]
      summary: 重建歌曲短语缓存 / Refresh song phrase cache
      operationId: refreshSongPhrases
      parameters: [{ $ref: '#/components/parameters/SongId' }]
      responses:
        '200': {description: 缓存已刷新}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/vocabulary/user-phrases:
    get:
      tags: [Phrases]
      summary: 查询用户短语 / List user phrases
      operationId: listUserPhrases
      responses:
        '200':
          description: 用户自定义短语
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/UserPhrase'}
    post:
      tags: [Phrases]
      summary: 添加用户短语 / Add a user phrase
      operationId: addUserPhrase
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserPhraseRequest'}
      responses:
        '200':
          description: 已保存的用户短语
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserPhrase'}
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/vocabulary/user-phrases/{id}:
    delete:
      tags: [Phrases]
      summary: 删除用户短语 / Delete a user phrase
      operationId: deleteUserPhrase
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        '200': {description: 已删除}
        '404': {$ref: '#/components/responses/NotFound'}

# ===============================================================
# Components
# ===============================================================
components:
  parameters:
    SongId:
      name: songId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: 请求参数错误 / Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40001
            message: page must be >= 0
            data: null
    NotFound:
      description: 资源未找到 / Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40401
            message: Word 'xyzabc' not found in vocabulary
            data: null
    InternalError:
      description: 服务器内部错误 / Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 50000
            message: Internal server error
            data: null
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 40001
        message:
          type: string
        data:
          nullable: true
      required: [code, message, data]
    VocabularyRebuildTask:
      type: object
      properties:
        taskId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, RUNNING, COMPLETED, FAILED]
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        errorMessage:
          type: string
          nullable: true
      required: [taskId, status, startedAt]
    WordPage:
      type: object
      properties:
        content:
          type: array
          items:
            type: string
        totalElements:
          type: integer
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer
      required: [content, totalElements, totalPages, number, size]
    WordOccurrence:
      type: object
      properties:
        songId:
          type: integer
          format: int64
        songTitle:
          type: string
          example: "Yesterday"
        songArtist:
          type: string
          example: "The Beatles"
        lyricLineId:
          type: integer
          format: int64
          nullable: true
        lineIndex:
          type: integer
        lyricLine:
          type: string
          example: "Yesterday, all my troubles seemed so far away"
        surfaceForm:
          type: string
          example: "running"
        lemma:
          type: string
          example: "run"
        startOffset:
          type: integer
        endOffset:
          type: integer
        learningScore:
          type: number
          format: double
      required: [songTitle, lyricLine]
    PhrasePage:
      type: object
      properties:
        content:
          type: array
          items: {$ref: '#/components/schemas/DictionaryPhrase'}
        totalElements: {type: integer}
        totalPages: {type: integer}
        number: {type: integer}
        size: {type: integer}
      required: [content, totalElements, totalPages, number, size]
    VocabularyQualityCandidate:
      type: object
      properties:
        word:
          type: string
        learningScore:
          type: number
          format: double
        occurrenceCount:
          type: integer
        songCount:
          type: integer
        recommended:
          type: boolean
        reasons:
          type: array
          items:
            type: string
            enum:
              - LOW_LEARNING_VALUE
              - PHRASE_CANDIDATE
              - CONTRACTION_PHRASE
              - POSSIBLE_TRUNCATED_LEMMA
              - VERY_SHORT_TOKEN
              - NON_STANDARD_TOKEN
        examples:
          type: array
          items:
            $ref: '#/components/schemas/WordOccurrence'
      required: [word, learningScore, occurrenceCount, songCount, recommended, reasons, examples]
    VocabularyBulkWordsRequest:
      type: object
      properties:
        words:
          type: array
          minItems: 1
          maxItems: 200
          items:
            type: string
      required: [words]
    VocabularyLearningValueRequest:
      type: object
      properties:
        recommended:
          type: boolean
      required: [recommended]
...
    DictionaryPhrase:
      type: object
      properties:
        id: {type: integer, format: int64}
        sourcePattern: {type: string}
        canonicalPattern: {type: string}
        definitionEn: {type: string}
        definitionZh: {type: string}
        usageNoteZh: {type: string}
        phraseType: {type: string}
        source: {type: string}
        tokenCountMin: {type: integer}
        tokenCountMax: {type: integer}
        matchPriority: {type: integer}
    PhraseMatch:
      type: object
      properties:
        phraseId: {type: integer, format: int64}
        sourcePattern: {type: string}
        canonicalPattern: {type: string}
        definitionEn: {type: string}
        definitionZh: {type: string}
        usageNoteZh: {type: string}
        phraseType: {type: string}
        source: {type: string}
        startTokenPosition: {type: integer}
        endTokenPosition: {type: integer}
        surfacePhrase: {type: string}
        matchPriority: {type: integer}
    PhraseOccurrence:
      type: object
      properties:
        phraseId: {type: integer, format: int64}
        songId: {type: integer, format: int64}
        songTitle: {type: string}
        songArtist: {type: string}
        lyricLineId: {type: integer, format: int64}
        lineIndex: {type: integer}
        lyricLine: {type: string}
        startTokenPosition: {type: integer}
        endTokenPosition: {type: integer}
        surfacePhrase: {type: string}
      required: [songTitle, lyricLine, surfacePhrase]
...
    UserPhraseRequest:
      type: object
      properties:
        canonicalPhrase: {type: string}
        definition: {type: string}
      required: [canonicalPhrase]
    UserPhrase:
      type: object
      properties:
        id: {type: integer, format: int64}
        userId: {type: string}
        canonicalPhrase: {type: string}
        definition: {type: string}
        createdAt: {type: string}
...
    TaskCreationResponse:
      type: object
      properties:
        taskId:
          type: string
          format: uuid
        message:
          type: string
      required: [taskId]